pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "afterburn"
version = "0.6.0"
description = "Post-training diagnostics for LLMs — what did fine-tuning actually do to your model?"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Mo", email = "mo@afterburn.dev" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "torch>=2.0,<3",
    "transformers>=4.35,<5",
    "safetensors>=0.4,<1",
    "numpy>=1.24,<3",
    "scipy>=1.10,<2",
    "click>=8.1,<9",
    "rich>=13.0,<14",
    "pyyaml>=6.0,<7",
    "jinja2>=3.1,<4",
    "plotly>=5.18,<6",
    "huggingface-hub>=0.19,<1",
]

[project.optional-dependencies]
pdf = ["weasyprint>=60,<61"]
semantic = ["sentence-transformers>=2.2,<3"]
nli = ["transformers>=4.35,<5"]
dev = [
    "afterburn[pdf]",
    "pytest>=7.4,<9",
    "pytest-cov>=4.1,<6",
    "hypothesis>=6.90,<7",
    "ruff>=0.1,<1",
    "mypy>=1.7,<2",
    "mkdocs-material>=9.0,<10",
]

[project.scripts]
afterburn = "afterburn.cli.main:cli"

[project.urls]
Homepage = "https://github.com/code-mohanprakash/afterburn"
Documentation = "https://github.com/code-mohanprakash/afterburn"
Repository = "https://github.com/code-mohanprakash/afterburn"

[tool.hatch.build.targets.wheel]
packages = ["src/afterburn"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (requiring model downloads)",
    "gpu: marks tests requiring GPU",
]
